Removed CallCustom.
CallCustom was added to provide a mechanism for binding a call site that was
independent of Ruby method resolution semantics.

The instruction set is being improved to, among other things, decouple method
or function resolution, caching, and invocation. This will provide arbitrary
semantics for finding the code to run, caching it, and invoking it.