Added more profiler functionality.
Use -Xsystem.profiler.target=<path> to set the output path for the profiler
report. The path can optionally contain the text $PID where desired and the
process ID will be substituted. Quoting the option may be necessary to prevent
the shell from expanding $PID prematurely. For example:

  rbx -Xsystem.profiler.target='path/to/profiler_report-$PID.txt'

The fully qualified name is implemented yet because it requires an additional
data structure to properly track the components, so only the bare method name
is displayed.