Removed support for private constants. Closes #3697.
Private constants are a bad idea, and supporting them isn't necessarily a good
idea.

Private constants do not add any capability to Ruby. They merely restrict
programs that would otherwise run. For this reason, not supporting private
constants in Rubinius doesn't prohibit any code that would otherwise run in
Ruby, except for the fruitless class of code that soley relies on accessing a
private constant to raise an exception.

The exception that is raised is a NameError, which causes confusion between a
constant not being defined at all, and therefore able to be defined or
autoloaded, etc., and the constant being defined but private, and therefore
*prohibiting* defining the constant. To distinguish between these cases would
require parsing the exception text itself.

Despite being labeled private and superficially not accessible, it is easy to
actually access the constant. So no well-meaning code is actually inhibited by
labeling a constant private.

If code were presented that relied on private constants being private, in the
positive sense (ie *added* capability, not merely prohibited capability), we
can revisit support for private constants.