Better backtrace rendering for copy-paste.
Rubinius renders backtraces with the receiver+method right justified and the
source location trailing. Previously, the source location was split when
exceeding the available terminal width to avoid cluttering the left side where
the receiver+method was rendered to aid visibility of the most important
backtrace information.

Unfortunately, splitting up the source location makes copy-pasting the path to
edit a source file very tedious.

Now, the source location is rendered as an unbroken sequence of characters but
starting on a new line if the total backtrace line exceeds the terminal width.
This preserves the readability of the receiver+method, while also preserving
the ability to copy-paste the source location.