fix each(limit) - longer explanation below
When calling each(limit) or each_line(limit) there is a lot of
logic to figure out if a separator or limit were passed. In
the case of #each, when a limit is specified it still picks up
$/ as the separator. Therefore EachReader#read_to_separator_with_limit
is used. There was a bug where a buffer was being assigned another
string fragment `buffer += result` which returned a new buffer
object. The calling method was relying on buffer being the same
object reference passed to its child method. Changing to #<<
resolved the bug.

Separately, the spec to test this behavior showed that #each_line
broke. It has similar logic to figure out separator or limit but
in this case it was passing each(5, 5, &blk) after figuring out
the limit. The fix is to reset sep_or_limit once we know what
it is supposed to be.

Incidently the only way to call EachReader#read_to_limit is to
set $/ = nil and send no separator in the #each method call.