Use unmanaged memory for CallSite.
This is a provisional implementation awaiting full support for reference
counting in the managed heap and garbage collector.

There are broadly four kinds of memory in Rubinius:

1. Memory on the call stack (ie not C locals) like the CallFrame object;
2. Unmanaged memory that is part of C/C++ structures (eg an instance of
   std::string) and really not closely related to managed (eg Ruby)
   objects;
3. Unmanaged memory that is closely related (eg MachineCode, CallSite)
   to managed (eg Ruby) objects;
4. Managed (eg Ruby) objects.

It's possible to put both 3 and 4 into the same category (ie managed
objects) if we allow for every one of those objects to use an extra
machine word or two (or some sort of side table) in order to keep track
of them.

The challenge in this approach is to ensure that the garbage collector
knows how to process them to find all the references they may contain.

By creating a composite model for the garbage collector that uses both
tracing and reference counting, it's possible to put managed objects in
places that won't be traced, and preserve their lifetime by counting how
many times they are referenced by live objects.

Another thing to note about 3 and 4 is that the difference mostly comes
down to whether the object will be exposed as a reference to the managed
(eg Ruby) code. For example, can it be stored in a local variable, or in
an Array or Hash? If it can be stored, then it needs to respond like a
managed object (eg what happens when it's passed to the method 'p').

This new mechanism maintains CallSite as a Ruby object because it's
useful for conveying the information at the call site to Ruby code. In
the future, arbitrary objects will be able to interact by means of
functions attached to them via FFI, so the system can handle non-managed
objects in a manner that is consistent with how it handles managed
objects.